+++ playback/src/player.rs
@@ -632,11 +632,11 @@ impl Drop for Player {
     }
 }
 
-struct PlayerLoadedTrackData {
-    decoder: Decoder,
+pub struct PlayerLoadedTrackData {
+    pub decoder: Decoder,
+    pub audio_item: AudioItem,
     normalisation_data: NormalisationData,
     stream_loader_controller: StreamLoaderController,
-    audio_item: AudioItem,
     bytes_per_second: usize,
     duration_ms: u32,
     stream_position_ms: u32,
@@ -655,7 +655,7 @@ enum PlayerPreload {
     },
 }
 
-type Decoder = Box<dyn AudioDecoder + Send>;
+pub type Decoder = Box<dyn AudioDecoder + Send>;
 
 enum PlayerState {
     Stopped,
@@ -876,9 +876,9 @@ impl PlayerState {
     }
 }
 
-struct PlayerTrackLoader {
-    session: Session,
-    config: PlayerConfig,
+pub struct PlayerTrackLoader {
+    pub session: Session,
+    pub config: PlayerConfig,
 }
 
 impl PlayerTrackLoader {
@@ -927,7 +927,7 @@ impl PlayerTrackLoader {
         Some(data_rate.ceil() as usize)
     }
 
-    async fn load_track(
+    pub async fn load_track(
         &self,
         spotify_id: SpotifyId,
         position_ms: u32,